frontend/requirements.txt
streamlit == 1.38.0
pandas == 2.2.3
streamlit-aggrid == 1.0.5
python-dotenv == 1.0.1
plotly == 5.24.1
pymongo == 3.12.0